requirements.txt
django
djangorestframework
django-filter
pillow
gunicorn
django-heroku
psycopg2-binary